requirements.txt
numpy
scipy
pandas
matplotlib
scikit-learn
pyyaml
torch
torchvision
gymnasium[box2d]
google-genai
tqdm
wandb
faiss-cpu
redis
prometheus-client
